- Fixed: ACS improperly calculated the address of local variables when   returning from one function to another function when the function that   was called was used as part of an expression. - Fixed: Using Thing_Hate with arg0 (hater) set to 0 from an open script   could crash. - Fixed: Some items along ledges in Hexen's MAP32 (Orchard of Lamentations)   appeared at the bottom of the ledge (and consequently inside it) instead   of on top of it because the items were placed directly on the lines.   AActor::LinkToWorldForMapThing() needs to use the original R_PointOnLineSide()   code to handle situations like this. Previously, it just used the original   code for straight horizontal/vertical lines and used the new code for   diagonal lines. - Fixed: FWadCollection::MergeLumps() used in incorrect realloc. - Fixed: FPlayList::NextLine() did not properly handle blank lines in the   playlist. - Changed: Decals now use lightweight thinkers instead of actors. (76 bytes   versus 396, so you save 320k if you have 1024 decals present.) - Fixed: Wads added with pullin were loaded immediately after the IWAD.   Exec files are now processed immediately before -file but after autoloading   wads in D_DoomMain(). - Fixed: sdl/i_system.h unconditionally defined SHARE_DIR, preventing   redefinition from the command line. - Fixed: The standard way to include SDL.h is <SDL.h>, not <SDL/SDL.h>. - Fixed: Returned FActiveInterpolation::HashKey()'s return type to size_t,   avoiding a pointer truncation warning.
SVN r30 (trunk)

+++ src/r_main.cpp
@@ -1774,9 +1774,9 @@ void FActiveInterpolation::DoAnInterpolation (fixed_t smoothratio)
 	*adr2 = oldipos[1] + FixedMul (pos - oldipos[1], smoothratio);
 }
 
-unsigned FActiveInterpolation::HashKey (EInterpType type, void *interptr)
+size_t FActiveInterpolation::HashKey (EInterpType type, void *interptr)
 {
-	return (unsigned)type * ((unsigned)interptr>>5);
+	return (size_t)type * ((size_t)interptr>>5);
 }
 
 int FActiveInterpolation::CountInterpolations ()
@@ -1823,7 +1823,7 @@ int FActiveInterpolation::CountInterpolations (int *usedbuckets, int *minbucketf
 
 FActiveInterpolation *FActiveInterpolation::FindInterpolation (EInterpType type, void *interptr, FActiveInterpolation **&interp_p)
 {
-	unsigned hash = HashKey (type, interptr) % INTERPOLATION_BUCKETS;
+	size_t hash = HashKey (type, interptr) % INTERPOLATION_BUCKETS;
 	FActiveInterpolation *probe, **probe_p;
 
 	for (probe_p = &curiposhash[hash], probe = *probe_p;
